Red repro: HEIC attachment silently kills the agent turn
Replays prod stream misha2 /agents/1648 in the unit lane: fixture is the
real journal up to offset 3007 (chunks + rawResponse stripped at dump
time), checkpoints primed to come into the chat halfway through, fake
Responses WebSocket behaves like real OpenAI (400 on the HEIC
input_image). Broad assertion: the user must still get an answer — red.

Extracts MemoryStream/deliverNewEvents/fakeResponsesWebSocket into
test-helpers.ts; MemoryStream.append now assigns last-offset+1 so seeded
histories with offset gaps don't collide.
